+++ R/children_of_parents.R
@@ -68,7 +68,13 @@ children_of_parents <- function(
 
 }
 
-#' @description Helper function for children_of_parents.
+#' Helper function for children_of_parents.
+#'
+#' @param parent_loc_ids [int] ihme location ids
+#' @param output [character] output options
+#' @param hierarchy [data.table] ihme location hierarchy
+#'
+#' @return [none] stop on failure
 validate_children_of_parents_inputs = function(parent_loc_ids, output, hierarchy){
   # Check for valid parent_locs_ids type
   if (!is.vector(parent_loc_ids) | !is.numeric(parent_loc_ids)) {
@@ -87,14 +93,15 @@ validate_children_of_parents_inputs = function(parent_loc_ids, output, hierarchy
   }
 }
 
-#' @description Helper function for children_of_parents.
+#' Helper function for children_of_parents.
+#'
 #' Given a single parent_id and a path_to_top_parent,
 #' returns TRUE if that parent_id is in the path.
 #'
 #' @param parent_id [int] Location ID of parent to test
 #' @param path_to_top_parent [character] String of path to top parent from hierarchy
 #'
-#' @return boolean
+#' @return [lgl] TRUE if parent_id is in path_to_top_parent
 is_child_of_parent = function(parent_id, path_to_top_parent){
   path_to_top_parent = as.integer(unlist(strsplit(path_to_top_parent, ",")))
   return(parent_id %in% path_to_top_parent)